support init BcosSDK with xml configuration files
1. create BcosSDK with configOption to support multiple configuration files

2. add applicationContext-sample.xml

3. Fix occasional write idle problem

+++ src/demo/java/org/fisco/bcos/sdk/demo/perf/collector/PerformanceCollector.java
@@ -130,7 +130,9 @@ public class PerformanceCollector {
                                     (double) (total - error.get()) / ((double) totalTime / 1000)));
             System.out.println("Avg time cost: " + String.valueOf(totalCost.get() / total) + "ms");
             System.out.println(
-                    "Error rate: " + String.valueOf((error.get() / received.get()) * 100) + "%");
+                    "Error rate: "
+                            + String.valueOf((error.get() / (double) received.get()) * 100)
+                            + "%");
 
             System.out.println("Time area:");
             System.out.println(
